No longer add msys2-keyring to IgnorePkg.
The official msys/i686 repository now contains the updated msys2-keyring
package.  The manual download/verify/install of the package is still
present, as the db is still signed by a new key, and even if it wasn't
pacman would likely fail due to mingw{32,64} being signed with a new
key.